Do not move header files that aren't mutated file configure file
