+++ clients/samples/10_gemm_amax_with_scale/sample_hipblaslt_gemm_amax_with_scale.cpp
@@ -25,7 +25,7 @@
  *******************************************************************************/
 
 #include <hip/hip_runtime.h>
-#include <hipblaslt-ext-op.h>
+#include <hipblaslt/hipblaslt-ext-op.h>
 #include <hipblaslt/hipblaslt.h>
 #include <iostream>
