+++ clients/samples/22_layernorm_ext/sample_hipblaslt_ext_op_layernorm.cpp
@@ -25,7 +25,7 @@
  *******************************************************************************/
 
 #include <hip/hip_runtime.h>
-#include <hipblaslt-ext-op.h>
+#include <hipblaslt/hipblaslt-ext-op.h>
 #include <hipblaslt/hipblaslt.h>
 #include <hipblaslt_arguments.hpp>
 #include <iostream>
