+++ clients/samples/23_amax_ext/sample_hipblaslt_ext_op_amax.cpp
@@ -25,7 +25,7 @@
  *******************************************************************************/
 
 #include <hip/hip_runtime.h>
-#include <hipblaslt-ext-op.h>
+#include <hipblaslt/hipblaslt-ext-op.h>
 #include <hipblaslt/hipblaslt.h>
 #include <iostream>
