+++ library/CMakeLists.txt
@@ -37,28 +37,20 @@ configure_file("${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt-version.h.in"
                "${PROJECT_BINARY_DIR}/include/hipblaslt/hipblaslt-version.h"
 )
 
-# Copy Public Headers to Build Dir
-configure_file( "${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt.h" "${PROJECT_BINARY_DIR}/include/hipblaslt/hipblaslt.h" COPYONLY)
-configure_file( "${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt-ext.hpp" "${PROJECT_BINARY_DIR}/include/hipblaslt/hipblaslt-ext.hpp" COPYONLY)
-configure_file( "${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt-types.h" "${PROJECT_BINARY_DIR}/include/hipblaslt/hipblaslt-types.h" COPYONLY)
-configure_file( "${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt_xfloat32.h" "${PROJECT_BINARY_DIR}/include/hipblaslt/hipblaslt_xfloat32.h" COPYONLY)
-configure_file( "${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt_float8.h" "${PROJECT_BINARY_DIR}/include/hipblaslt/hipblaslt_float8.h" COPYONLY)
-configure_file( "${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt_float8_bc.h" "${PROJECT_BINARY_DIR}/include/hipblaslt/hipblaslt_float8_bc.h" COPYONLY)
-configure_file( "${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt-ext-op.h" "${PROJECT_BINARY_DIR}/include/hipblaslt/hipblaslt-ext-op.h" COPYONLY)
 
 # Public hipBLASLt headers
 set(hipblaslt_headers_public
-  include/hipblaslt.h
-  include/hipblaslt-ext.hpp
-  include/hipblaslt_xfloat32.h
-  include/hipblaslt_float8.h
-  include/hipblaslt_float8_bc.h
-  include/hipblaslt-ext-op.h
+  include/hipblaslt/hipblaslt.h
+  include/hipblaslt/hipblaslt-ext.hpp
+  include/hipblaslt/hipblaslt_xfloat32.h
+  include/hipblaslt/hipblaslt_float8.h
+  include/hipblaslt/hipblaslt_float8_bc.h
+  include/hipblaslt/hipblaslt-ext-op.h
   ${PROJECT_BINARY_DIR}/include/hipblaslt/hipblaslt-version.h
 )
 
 if(NOT BUILD_CUDA)
-    set(hipblaslt_headers_public ${hipblaslt_headers_public} include/hipblaslt-types.h)
+    set(hipblaslt_headers_public ${hipblaslt_headers_public} include/hipblaslt/hipblaslt-types.h)
 endif()
 source_group("Header Files\\Public" FILES ${hipblaslt_headers_public})
 
@@ -126,7 +118,6 @@ target_include_directories(hipblaslt
                            PRIVATE $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src/include>
                                    $<BUILD_INTERFACE:${Tensile_INC}>
                            PUBLIC  $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/library/include>
-                                   $<BUILD_INTERFACE:${PROJECT_BINARY_DIR}/include>
                                    $<BUILD_INTERFACE:${HIP_INCLUDE_DIRS}>
                                    $<INSTALL_INTERFACE:include>
 )
@@ -222,6 +213,12 @@ rocm_install_targets(TARGETS hipblaslt
                      INCLUDE
                        ${CMAKE_BINARY_DIR}/include
 )
+install(
+    DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt"
+    DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}"
+    PATTERN "*.h"
+    PATTERN "*.hpp"
+)
 
 if ( NOT BUILD_CUDA )
     if (WIN32)
@@ -264,9 +261,9 @@ endif()
 
 
 if(BUILD_FILE_REORG_BACKWARD_COMPATIBILITY)
-  rocm_install(
-    DIRECTORY
-       "${PROJECT_BINARY_DIR}/hipblaslt"
-        DESTINATION "." )
-  message( STATUS "Backward Compatible Sym Link Created for include directories" )
+rocm_install(
+  DIRECTORY
+     "${CMAKE_CURRENT_SOURCE_DIR}/include/hipblaslt"
+      DESTINATION "." )
+message( STATUS "Backward Compatible Sym Link Created for include directories" )
 endif()
